+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Testing:
     * Align with new Zenodo API (\#601).
     * Update `test_valid_manifest` (\#606).
+    * Use [pooch](https://www.fatiando.org/pooch) to download test files (\#610).
 
 # 0.13.1